% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularBar.R
\name{circularBar}
\alias{circularBar}
\title{circularBar}
\usage{
circularBar(df_edges, dict.combine, ColorsCirc)
}
\arguments{
\item{df_edges}{dataframe. "from", "to", "weight".}

\item{dict.combine}{dataframe. "id", "label", "term", "semantic_type", "group1", "group", "type", "category"}

\item{ColorsCirc}{dataframe. "", "", ""}
}
\value{
a circular bar plot.
}
\description{
Plot a bar plot with polar coordinate.
}
\examples{
\dontrun{
sunburstPlotly("node is", "df_edges", "dict.combine")
}
}
