% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clickedNodeText.R
\name{clickedNodeText}
\alias{clickedNodeText}
\title{clicked node text}
\usage{
clickedNodeText(node_id, dict.combine)
}
\arguments{
\item{node_id}{string. clicked node id.}

\item{dict.combine}{data.frame. "id", "semantic_type", "category"}
}
\value{
sets of tags.
}
\description{
Module for header
}
\examples{
\dontrun{
clickedNodeText(node_id, dict.combine)
}
}
