% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailsTab.R
\name{detailsTab}
\alias{detailsTab}
\title{details tab}
\usage{
detailsTab(tname, df, title, outdiv, nodeid, output, has_sy, helps = "")
}
\arguments{
\item{tname}{string. name of the details data.}

\item{df}{data.frame. details data.}

\item{title}{data.frame. "id", "label", "term", "semantic_type", "group2", "group", "type", "category"}

\item{outdiv}{list. tagList.}

\item{nodeid}{string. node id.}

\item{output}{output of shiny server.}

\item{has_sy}{boolean. Default FALSE. synonyms table.}

\item{helps}{string. Default "". Description about the data.}
}
\value{
a visNetwork plot.
}
\description{
show a table or list for more details about the selected node.
}
\examples{
\dontrun{
plot_network("df_edges", "dict.combine", "attrs", "colors_group")
}
}
