% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_DB.R
\name{getData}
\alias{getData}
\alias{data2db}
\alias{writeTable}
\title{Get/write data from a PostgreSQL database}
\usage{
getData(id, tname, db, field = "from")

data2db(data, tname, db, title = NA, note = NA)

writeTable(data, tname, db)
}
\arguments{
\item{id}{string. The id of the data to get from the table. When NULL, the entire table will be returned.}

\item{tname}{string. The table name of the data.  A table name can contain numeric alphabets and underline (e.g. table_1 ). The table name  should begin with an alphabet. The 8 table names below are specified:
* "edge": matrix data for network.
* "node": dictionary of the nodes.
* "synonyms": synonyms of CUI nodes.
* "rollup": rollup for CUI nodes.
* some other names can't be used: "details".}

\item{db}{string. The name of the database.}

\item{field}{string. Default from. The field name of the id to get from the table.}

\item{data}{data.frame. The data to write into the database.}

\item{title}{string. Default NULL. Parameter for more data. The title for more information data.}

\item{note}{string. Default "". Parameter for more data. The description for more information data.}
}
\description{
getData: get the entire table or some rows from the table from a PostgreSQL database.
writeTable: write the table into a PostgreSQL database.
}
\examples{
\dontrun{
getData("PheCode:250", "node", "app_cuinetwork.db")
data2db("edge", "test_data2db.db", "test_data2db.rds")
}
}
