% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header.R
\name{headerServer}
\alias{headerServer}
\title{Server Module for header}
\usage{
headerServer(id, steps, doc, data)
}
\arguments{
\item{id}{string. Namespace of the module.}

\item{steps}{dataframe. The step-by-step introduction.}

\item{doc}{string. The path of the doc.}

\item{data}{dataframe. The data to download.}
}
\value{
server part for header UI.
}
\description{
Server Module for header
}
\examples{
\dontrun{
headerServer('header')
}
}
