% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header.R
\name{headerUI}
\alias{headerUI}
\title{UI Module for header}
\usage{
headerUI(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
sets of tags.
}
\description{
Module for header
}
\examples{
\dontrun{
headerUI('header')
}
}
