% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{plot_network}
\alias{plot_network}
\title{plot network}
\usage{
plot_network(
  center_nodes,
  df_edges,
  dict.combine,
  attrs,
  colors_group,
  shapes,
  hide_labels = TRUE,
  directed = FALSE,
  node_num_cutoff = 500,
  layout = "layout_nicely"
)
}
\arguments{
\item{center_nodes}{vector. center nodes' id.}

\item{df_edges}{data.frame. "from", "to", "weight".}

\item{dict.combine}{data.frame. "id", "label", "term", "semantic_type", "group2", "group", "type", "category"}

\item{attrs}{list. attr_nodes_type attr_nodes_center attr_edges attr_legend_edge(if directed)}

\item{colors_group}{data.frame. "group", "color.background".}

\item{shapes}{data.frame. "type", "shape".}

\item{hide_labels}{logical. Default TRUE.}

\item{directed}{logical. Default FALSE.}

\item{node_num_cutoff}{numeric. Default 500. Most nodes to show.}

\item{layout}{string. Default "layout_nicely". layout for igraph.}
}
\value{
a visNetwork plot.
}
\description{
Plot the network.
}
\examples{
\dontrun{
plot_network("df_edges", "dict.combine", "attrs", "colors_group")
}
}
