% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shapePicker.R
\name{shapepickerUI}
\alias{shapepickerUI}
\alias{shapepickerServer}
\title{mod_shapepicker}
\usage{
shapepickerUI(id, shape)

shapepickerServer(id)
}
\arguments{
\item{id}{string. type of the network node.}

\item{shape}{string. "square", "dot", "triangle", "star", "box", "circle",
"ellipse", "database", "text", "diamond".}
}
\value{
shapepickerUI returns a pickerInput of shape for different types.
        shapepickerServer returns the value of shape pickerInput.
}
\description{
module of shapepicker.
}
\examples{
\dontrun{
shapepickerUI("NLP", "square")
shapepickerServer("NLP")
}
}
