% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{sidebarUI}
\alias{sidebarUI}
\alias{sidebarServer}
\title{Module for sidebar}
\usage{
sidebarUI(id)

sidebarServer(
  id,
  tb_input,
  tname,
  db,
  type = 1,
  selected = c(1, 3),
  init_nodes = NULL,
  server = TRUE
)
}
\arguments{
\item{id}{string. Namespace of the module.}

\item{tb_input}{string. Namespace of the module.}

\item{tname}{string. table name in db to search.}

\item{db}{string. name of the database.}

\item{type}{string. Default DT. DT or reactable.}

\item{selected}{numeric. Pre-selected rows in table. Default c(1,3).}

\item{init_nodes}{vector. Default NULL. Initial center nodes.}

\item{server}{logical. Default TRUE, the data is kept on the server and
the browser requests a page at a time; if FALSE, then the entire 
data frame is sent to the browser at once. Highly recommended for 
medium to large data frames, which can cause browsers to slow down or 
crash. Passed to renderDT().}
}
\value{
sets of tags.

vector of center nodes.
}
\description{
Module for sidebar
}
\examples{
\dontrun{
sidebarUI('input')
}
\dontrun{
sidebarServer('input')
}
}
