% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburstPlotly}
\alias{sunburstPlotly}
\title{plot sunburst}
\usage{
sunburstPlotly(node_now, df_edges, dict.combine)
}
\arguments{
\item{node_now}{string. the selected node's id.}

\item{df_edges}{dataframe. "from", "to", "weight".}

\item{dict.combine}{dataframe. "id", "label", "term", "semantic_type", "group2", "group", "type", "category"}
}
\value{
a sunburst plot.
}
\description{
Plot the sunburst.
}
\examples{
\dontrun{
sunburstPlotly("node is", "df_edges", "dict.combine")
}
}
