% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tbClicked.R
\name{tbClickedUI}
\alias{tbClickedUI}
\title{UI Module for reactable of clicked node}
\usage{
tbClickedUI(id)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}
}
\value{
sets of tags.
}
\description{
Module for reactable of clicked node
}
\examples{
\dontrun{
tbClickedUI('tb_clicked')
}
}
