% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_windowSize.R
\name{windowSizeServer}
\alias{windowSizeServer}
\title{Server Module for window size}
\usage{
windowSizeServer(id)
}
\arguments{
\item{id}{string. namespace the module}
}
\value{
a numeric vector of height and width of the window.
}
\description{
Server Module for using JS and get the size of the window refer to https://stackoverflow.com/questions/47424498/utilizing-window-size-within-shiny-module.
}
\examples{
\dontrun{
windowSizeServer('body')
}
}
