% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_windowSize.R
\name{windowSizeUI}
\alias{windowSizeUI}
\title{UI Module for window size}
\usage{
windowSizeUI(id)
}
\arguments{
\item{id}{string. namespace the module}
}
\value{
JS in UI.
}
\description{
Module for using JS and get the size of the window refer to https://stackoverflow.com/questions/47424498/utilizing-window-size-within-shiny-module.
}
\examples{
\dontrun{
windowSizeUI('body')
}
}
