% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_DB.R
\name{getData}
\alias{getData}
\alias{data2db}
\alias{buildIndex}
\title{Get/write data from a PostgreSQL database}
\usage{
getData(id, tname, db, field = "from")

data2db(data, tname, db, title = NA, note = NA)

buildIndex(field, tname, db)
}
\arguments{
\item{id}{string. The id of the data to get from the table. When NULL, the entire table will be returned.}

\item{tname}{string. The table name of the data.  A table name can contain numeric alphabets and underline (e.g. table_1 ). The table name  should begin with an alphabet. 
The 3 table names below are specified: "df_edges": data for network; "dict": dictionary of the nodes; "details".}

\item{db}{string. The name of the database.}

\item{field}{string. Default "from". The field name of the id to get from the table.}

\item{data}{data.frame. The data to write into the database.}

\item{title}{string. Default NULL. Parameter for more data. The title for more information data.}

\item{note}{string. Default "". Parameter for more data. The description for more information data.}
}
\description{
getData: get the entire table or some rows from the table from a PostgreSQL database. 

data2db: write the table into a PostgreSQL database. 

buildIndex: build index for a table
}
\examples{
\dontrun{
getData("PheCode:250", "df_edges", "app_cuinetwork")
data2db(df_edges, "df_edges", "app_cuinetwork")
}
}
